requirements.txt
Django>=4.2
django-cors-headers>=4.3
whitenoise>=6.6
gunicorn>=21.2
dj-database-url>=2.1
psycopg2-binary>=2.9
djangorestframework>=3.14  # Si vous utilisez DRF